Console: move tests to the correct package
